+++ src/estd/optional.h
@@ -11,7 +11,6 @@
 #include "memory.h"
 
 #include "internal/feature/optional.h"
-#include "internal/fwd/functional.h"
 #include "internal/functional/hash.h"
 #include "internal/optional.h"
 
@@ -409,19 +408,26 @@ public:
 
 }
 
-template <class T, class U, class TBase>
-ESTD_CPP_CONSTEXPR_RET bool operator==(const optional<T, TBase>& opt, const U& value)
+template <class T, class U, class Base>
+constexpr bool operator==(const optional<T, Base>& opt, const U& value)
 {
     return opt.has_value() ? *opt == value : false;
 }
 
 
-template <class T, class U, class TBase>
-ESTD_CPP_CONSTEXPR_RET bool operator==(const U& value, const optional<T, TBase>& opt)
+template <class T, class U, class Base>
+constexpr bool operator==(const U& value, const optional<T, Base>& opt)
 {
     return opt.has_value() ? value == *opt : false;
 }
 
+template <class T, class T2, class Base, class Base2>
+constexpr bool operator==(const optional<T, Base>& lhs, const optional<T2, Base2>& rhs)
+{
+    return lhs.has_value() && rhs.has_value() ?
+        *lhs == *rhs :
+        !lhs.has_value() && !rhs.has_value();
+}
 
 template <class T, class U, class TBase>
 ESTD_CPP_CONSTEXPR_RET bool operator!=(const optional<T, TBase>& opt, const U& value)
@@ -460,9 +466,9 @@ struct hash<optional<T, Base>>
 {
     constexpr size_t operator()(const optional<T, Base>& v) const
     {
-        using underlying = hash<T>;
+        using underlying = hash<estd::remove_const_t<T>>;
 
-        return v ? underlying{}(*v) : 0;
+        return underlying{}(*v);
     }
 };